Comments on Final Project
This is the 12th version of my stock tracker app.  This is my final project.  I wanted to do more, but i don't think I'll be able to within the amount of time.  This app uses a python module called pandas to run correlations on historical data between the Dow Jones Industrial average, the Nasdaq, and the S&P500.  The majority of the code, however, is devoted to creating a new csv file called 'Master.csv' which feeds the pandas module's correlation function.